+++ broker/src/asapo_broker/server/server.go
@@ -2,14 +2,15 @@ package server
 
 import (
 	"asapo_broker/database"
+	"asapo_common/discovery"
 	log "asapo_common/logger"
 	"errors"
-	"io/ioutil"
 	"net/http"
 )
 
 const kDefaultresendInterval = 10
 const kDefaultStreamCacheUpdateIntervalMs = 100
+const kDefaultTokenCacheUpdateIntervalMs = 60000
 
 var db database.Agent
 
@@ -25,6 +26,14 @@ type serverSettings struct {
 	discoveredDbAddress         string
 	CheckResendInterval         *int
 	StreamCacheUpdateIntervalMs *int
+	TokenCacheUpdateIntervalMs  *int
+}
+
+func (s *serverSettings) GetTokenCacheUpdateInterval() int {
+	if s.TokenCacheUpdateIntervalMs == nil {
+		return kDefaultTokenCacheUpdateIntervalMs
+	}
+	return *s.TokenCacheUpdateIntervalMs
 }
 
 func (s *serverSettings) GetResendInterval() int {
@@ -58,17 +67,7 @@ type discoveryAPI struct {
 	baseURL string
 }
 
-var discoveryService discoveryAPI
-
-func (api *discoveryAPI) GetMongoDbAddress() (string, error) {
-	resp, err := api.Client.Get(api.baseURL + "/asapo-mongodb")
-	if err != nil {
-		return "", err
-	}
-	defer resp.Body.Close()
-	body, err := ioutil.ReadAll(resp.Body)
-	return string(body), err
-}
+var discoveryService discovery.DiscoveryAPI
 
 func ReconnectDb() (err error) {
 	if db == nil {
@@ -97,7 +96,7 @@ func InitDB(dbAgent database.Agent) (err error) {
 }
 
 func CreateDiscoveryService() {
-	discoveryService = discoveryAPI{&http.Client{}, "http://" + settings.DiscoveryServer}
+	discoveryService = discovery.CreateDiscoveryService(&http.Client{},"http://" + settings.DiscoveryServer)
 }
 
 func CleanupDB() {