+++ discovery/src/asapo_discovery/protocols/hard_coded_producer.go
@@ -2,11 +2,16 @@ package protocols
 
 func GetSupportedProducerProtocols() []Protocol {
 	return []Protocol{
+		Protocol{"v0.4",
+			map[string]string{
+				"Discovery": "v0.1",
+				"Receiver": "v0.4",
+			}, &protocolValidatorCurrent{}},
 		Protocol{"v0.3",
 			map[string]string{
 				"Discovery": "v0.1",
 				"Receiver": "v0.3",
-			}, &protocolValidatorCurrent{}},
+			}, &protocolValidatorDeprecated{getTimefromDate("2022-09-01")}},
 		Protocol{"v0.2",
 			map[string]string{
 				"Discovery": "v0.1",